feat(modeling): add focal loss-star
